backend/requirements.txt
# FaceAttend-KE Backend Dependencies
# Optimized for Render.com deployment (Linux/PostgreSQL)
# Python 3.11+ Required

# =============================================================================
# CORE FRAMEWORK
# =============================================================================
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
python-dotenv==1.0.1

# =============================================================================
# DATABASE - PostgreSQL for Render (replaces MySQL)
# =============================================================================
# psycopg2-binary for PostgreSQL on Render
psycopg2-binary==2.9.9

# SQLAlchemy core (required by Flask-SQLAlchemy)
SQLAlchemy==2.0.23

# =============================================================================
# FACE RECOGNITION - Linux-compatible builds
# =============================================================================
# face-recognition and dependencies (Render runs on Linux)
face-recognition==1.3.0
dlib==19.24.2
opencv-python-headless==4.9.0.80
numpy==1.24.3
Pillow==10.2.0
scipy==1.11.4

# =============================================================================
# SECURITY
# =============================================================================
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0

# =============================================================================
# UTILITIES
# =============================================================================
pytz==2024.1
python-dateutil==2.8.2
requests==2.31.0

# =============================================================================
# PRODUCTION SERVER (Render uses gunicorn on Linux)
# =============================================================================
gunicorn==21.2.0

# =============================================================================
# DEVELOPMENT (optional - not installed on Render by default)
# =============================================================================
# flask-debugtoolbar==0.14.1
# pytest==7.4.3
# pytest-cov==4.1.0